Remove delete button if user is admin

+++ resources/views/admin/users/show.blade.php
@@ -8,18 +8,20 @@
 
             <a href="{{ url('/admin/users') }}" title="Back"><button class="btn btn-secondary btn-sm"><i class="fa fa-arrow-left" aria-hidden="true"></i> Back</button></a>
             <a href="{{ url('/admin/users/' . $user->id . '/edit') }}" title="Edit User"><button class="btn btn-primary btn-sm"><i class="fa fa-edit" aria-hidden="true"></i> Edit</button></a>
-            {!! Form::open([
-                'method' => 'DELETE',
-                'url' => ['/admin/users', $user->id],
-                'style' => 'display:inline'
-            ]) !!}
+            @if (!$user->role()->first() === 'admin')
+                {!! Form::open([
+                    'method' => 'DELETE',
+                    'url' => ['/admin/users', $user->id],
+                    'style' => 'display:inline'
+               ]) !!}
                 {!! Form::button('<i class="fa fa-trash" aria-hidden="true"></i> Delete', array(
                         'type' => 'submit',
                         'class' => 'btn btn-danger btn-sm',
                         'title' => 'Delete User',
                         'onclick'=>'return confirm("Confirm delete?")'
                 ))!!}
-            {!! Form::close() !!}
+                {!! Form::close() !!}
+            @endif
             <br/>
             <br/>
